Running Eslint against test code
Up to now, we were not running lint on any test file.
This commit applies lint rules to test files and fixes the problems that
arose from this.
