+++ .eslintignore
@@ -1,2 +1,1 @@
-test/
 node_modules/
