+++ .eslintrc.js
@@ -7,7 +7,7 @@ module.exports = {
   'rules': {
     'extra-rules/no-commented-out-code': 2,
     'extra-rules/no-for-loops': 2,
-    'extra-rules/no-long-files': [2, 113],
+    'extra-rules/no-long-files': [2, 195],
     'extra-rules/potential-point-free': 2
   },
   'parserOptions': {
