+++ package.json
@@ -7,7 +7,7 @@
     "start": "babel-node main.js",
     "start-dev": "nodemon main.js --exec babel-node",
     "test": "snyk test && npm run lint && npm run unit-tests",
-    "unit-tests": "NODE_ENV=test mocha 'test/**/*_spec.js'",
+    "unit-tests": "NODE_ENV=test mocha 'test/**/*_spec.js' --require test/index.js",
     "lint": "eslint '**/*.js'"
   },
   "keywords": [
