+++ test/controllers/equipment/options_equipment_spec.js
@@ -1,0 +1,38 @@
+import app from '../../../app';
+import CanVariablesFetcher from '../../../fetcher/canVariablesFetcher';
+import EquipmentFetcher from '../../../fetcher/equipmentFetcher';
+import td from 'testdouble';
+import TrackingPointFetcher from '../../../fetcher/trackingPointFetcher';
+
+describe('EquipmentController', () => {
+  let canVariablesFetcher;
+  let equipmentFetcher;
+  let trackingPointFetcher;
+  let server;
+
+  beforeEach(() => {
+    canVariablesFetcher = td.object(CanVariablesFetcher);
+    equipmentFetcher = td.object(EquipmentFetcher);
+    trackingPointFetcher = td.object(TrackingPointFetcher);
+
+    server = app(equipmentFetcher, canVariablesFetcher, trackingPointFetcher);
+  });
+
+  describe('OPTIONS /equipment', () => {
+    it('should return the list of allowed HTTP methods', (done) => {
+      const optionsRequest = {
+        url: '/equipment',
+        method: 'OPTIONS'
+      };
+
+      const expectedResponse = '';
+
+      server.inject(optionsRequest, (res) => {
+        expect(res.statusCode).to.be.eql(200);
+        expect(res.headers.allow).to.be.eql('GET');
+        expect(res.payload).to.be.eql(expectedResponse);
+        done();
+      });
+    });
+  });
+});
