+++ test/index.js
@@ -1,0 +1,3 @@
+import chai from 'chai';
+
+global.expect = chai.expect;
